refactor: split vector-stores into flat file modules
Split each monolithic src/index.ts (220-364 lines) into focused modules:
- types.ts (client interfaces, config types)
- store.ts (*VectorStore class)
- search-index.ts (*VectorIndex class)
- filters.ts (VectorFilter translation)
- helpers.ts (serialization, parsing, ID generation)
- index.ts (barrel re-exports)

Applied to all 5 packages: chroma, qdrant, pgvector, pinecone, milvus.
Public API surface unchanged -- all exports preserved via barrel.
