+++ packages/vector-stores/chroma/src/search-index.ts
@@ -1,0 +1,40 @@
+import { type EmbeddingModel, embedText, type VectorMetadata } from "@anvia/core/embeddings";
+import type { Tool } from "@anvia/core/tool";
+import {
+  createVectorSearchTool,
+  type VectorSearchIndex,
+  type VectorSearchRequest,
+  type VectorSearchResult,
+  type VectorSearchToolOptions,
+} from "@anvia/core/vector-store";
+import { filterToChromaWhere } from "./filters.js";
+import { parseQueryResults } from "./helpers.js";
+import type { ChromaCollectionLike } from "./types.js";
+
+export class ChromaVectorIndex<T, Metadata extends VectorMetadata = VectorMetadata>
+  implements VectorSearchIndex<T, Metadata>
+{
+  constructor(
+    private readonly model: EmbeddingModel,
+    private readonly collection: ChromaCollectionLike,
+  ) {}
+
+  async search(request: VectorSearchRequest): Promise<Array<VectorSearchResult<T, Metadata>>> {
+    const queryEmbedding = await embedText(this.model, request.query);
+    const response = await this.collection.query({
+      queryEmbeddings: [queryEmbedding.vector],
+      nResults: request.topK,
+      where: filterToChromaWhere(request.filter),
+      include: ["documents", "metadatas", "distances"],
+    });
+    return parseQueryResults<T, Metadata>(response, request.threshold);
+  }
+
+  async searchIds(request: VectorSearchRequest): Promise<Array<{ score: number; id: string }>> {
+    return (await this.search(request)).map(({ score, id }) => ({ score, id }));
+  }
+
+  asTool(options: VectorSearchToolOptions): Tool<{ query: string; topK?: number }, unknown> {
+    return createVectorSearchTool(this, options);
+  }
+}
