+++ packages/vector-stores/qdrant/src/search-index.ts
@@ -1,0 +1,41 @@
+import { type EmbeddingModel, embedText, type VectorMetadata } from "@anvia/core/embeddings";
+import type { Tool } from "@anvia/core/tool";
+import {
+  createVectorSearchTool,
+  type VectorSearchIndex,
+  type VectorSearchRequest,
+  type VectorSearchResult,
+  type VectorSearchToolOptions,
+} from "@anvia/core/vector-store";
+import { filterToQdrantFilter } from "./filters.js";
+import { parseQueryResults } from "./helpers.js";
+import type { QdrantClientLike } from "./types.js";
+
+export class QdrantVectorIndex<T, Metadata extends VectorMetadata = VectorMetadata>
+  implements VectorSearchIndex<T, Metadata>
+{
+  constructor(
+    private readonly model: EmbeddingModel,
+    private readonly client: QdrantClientLike,
+    private readonly collectionName: string,
+  ) {}
+
+  async search(request: VectorSearchRequest): Promise<Array<VectorSearchResult<T, Metadata>>> {
+    const queryEmbedding = await embedText(this.model, request.query);
+    const response = await this.client.search(this.collectionName, {
+      vector: queryEmbedding.vector,
+      limit: request.topK,
+      filter: filterToQdrantFilter(request.filter),
+      with_payload: true,
+    });
+    return parseQueryResults<T, Metadata>(response, request.threshold);
+  }
+
+  async searchIds(request: VectorSearchRequest): Promise<Array<{ score: number; id: string }>> {
+    return (await this.search(request)).map(({ score, id }) => ({ score, id }));
+  }
+
+  asTool(options: VectorSearchToolOptions): Tool<{ query: string; topK?: number }, unknown> {
+    return createVectorSearchTool(this, options);
+  }
+}
